test: 장바구니 통합 테스트를 기능 단위로 파일 분리
